+++ src/TestSuite/TestCase.php
@@ -15,9 +15,11 @@ namespace Thumber\Cake\TestSuite;
 
 use Cake\TestSuite\TestCase as CakeTestCase;
 use Thumber\Cake\Utility\ThumbCreator;
-use Thumber\TestSuite\TestTrait as ThumberTestTrait;
-use Tools\ReflectionTrait;
+//use Thumber\TestSuite\TestTrait as ThumberTestTrait;
+use Tools\TestSuite\ReflectionTrait;
 use Tools\TestSuite\TestTrait as ToolsTestTrait;
+use Thumber\TestSuite\TestTrait;
+use Tools\TestSuite\BackwardCompatibilityTrait;
 
 /**
  * TestCase class
@@ -25,8 +27,10 @@ use Tools\TestSuite\TestTrait as ToolsTestTrait;
 abstract class TestCase extends CakeTestCase
 {
     use ReflectionTrait;
-    use ThumberTestTrait;
+//    use ThumberTestTrait;
     use ToolsTestTrait;
+    use BackwardCompatibilityTrait;
+    use TestTrait;
 
     /**
      * Asserts for a valid thumbnail url
@@ -38,19 +42,36 @@ abstract class TestCase extends CakeTestCase
      */
     public function assertThumbUrl($url, $message = '')
     {
-        self::assertRegExp('/^(http:\/\/localhost)?\/thumb\/[\w\d]+/', $url, $message);
+        self::assertMatchesRegularExpression('/^(http:\/\/localhost)?\/thumb\/[\w\d]+/', $url, $message);
     }
 
     /**
      * Returns an instance of `ThumbCreator`
-     * @param string|null $path Path of the image from which to create the
+     * @param string $path Path of the image from which to create the
      *  thumbnail. It can be a relative path (to APP/webroot/img), a full path
      *  or a remote url
      * @return \Thumber\Cake\Utility\ThumbCreator
-     * @since 1.5.1
      */
-    protected function getThumbCreatorInstance($path = null)
+    protected function getThumbCreatorInstance($path = '')
     {
         return new ThumbCreator($path ?: '400x400.jpg');
+    }
+
+    /**
+     * Returns an instance of `ThumbCreator`, after calling `resize()` and
+     *  `save()` methods.
+     *
+     * It can be called passing only the array of options as first argument.
+     * @param string $path Path of the image from which to create the thumbnail.
+     *  It can be a full path or a remote url
+     * @param array $options Options for saving
+     * @return \Thumber\Cake\Utility\ThumbCreator
+     */
+    protected function getThumbCreatorInstanceWithSave(string $path = '', array $options = []): ThumbCreator
+    {
+        $thumbCreator = $this->getThumbCreatorInstance($path);
+        $thumbCreator->resize(200)->save($options);
+
+        return $thumbCreator;
     }
 }